+++ .build/tasks/Invoke-Pester.pester.build.ps1
@@ -67,7 +67,7 @@ task Invoke_pester_tests {
         ProjectName     = $ProjectName
     }
 
-    $ModuleVersion = Get-ModuleVersion @getModuleVersionParameters
+    $ModuleVersion = Get-BuiltModuleVersion @getModuleVersionParameters
 
     if (!(Test-Path $PesterOutputFolder))
     {
@@ -306,7 +306,7 @@ task Fail_Build_if_Pester_Tests_failed {
         ProjectName     = $ProjectName
     }
 
-    $ModuleVersion = Get-ModuleVersion @GetModuleVersionParameters
+    $ModuleVersion = Get-BuiltModuleVersion @GetModuleVersionParameters
 
     $powerShellVersion = 'PSv.{0}' -f $PSVersionTable.PSVersion
     $getPesterOutputFileFileNameParameters = @{
@@ -375,7 +375,7 @@ task Pester_if_Code_Coverage_Under_Threshold {
         ProjectName     = $ProjectName
     }
 
-    $ModuleVersion = Get-ModuleVersion @GetModuleVersionParameters
+    $ModuleVersion = Get-BuiltModuleVersion @GetModuleVersionParameters
 
     $powerShellVersion = 'PSv.{0}' -f $PSVersionTable.PSVersion
 
@@ -450,7 +450,7 @@ task Upload_Test_Results_To_AppVeyor -If { (property BuildSystem 'unknown') -eq 
         ProjectName     = $ProjectName
     }
 
-    $ModuleVersion = Get-ModuleVersion @GetModuleVersionParameters
+    $ModuleVersion = Get-BuiltModuleVersion @GetModuleVersionParameters
 
     $powerShellVersion = 'PSv.{0}' -f $PSVersionTable.PSVersion